further refactor to components
